requirements.txt
# Updated requirements for DeepV (2026)
# Pinned to compatible modern versions for stability
# Install with: pip install -r requirements.txt

# Core ML
torch==2.5.1
torchvision==0.20.1
numpy==1.24.3
scipy==1.10.1
scikit-learn==1.2.2

# Image processing
Pillow==9.5.0
opencv-python==4.8.1.78
scikit-image==0.21.0

# Rendering and vector graphics
cairocffi==1.7.1
svgpathtools==1.7.2
svgwrite==1.4.3
reportlab==4.0.4
css-parser==1.0.9

# Utilities
tqdm==4.65.0
tensorboardx==2.6
matplotlib==3.7.2
plotly==5.15.0

# Config and logging
omegaconf==2.3.0
hydra-core==1.3.2
wandb==0.15.12

# Other deps (minimal subset from original)
absl-py==1.4.0
protobuf==3.20.3
requests==2.31.0
pyyaml==6.0.1
h5py==3.9.0
Cython==3.0.0
IPython==8.12.0
rtree==1.0.1

# Note: For GPU support, install torch/torchvision with CUDA:
# pip install torch torchvision --index-url https://download.pytorch.org/whl/cu121